requirements.txt
# FAP - Field Audit Platform Dependencies

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0

# Testing & Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data Processing & RF Analysis
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.8.0
pandas>=2.0.0

# Report Generation
reportlab>=4.0.0
python-docx>=1.1.0
jinja2>=3.1.0

# Device Integration
paho-mqtt>=1.6.0
pyserial>=3.5

# Development Tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Environment & Config
python-dotenv>=1.0.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0